Move PackageFile class to its own module
